+++ libs/angular/.storybook/postcss.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  // Add your installed PostCSS plugins here:
-  plugins: [require('autoprefixer')],
-}