+++ libs/core/src/components/filter-chips/filter-chips.styles.ts
@@ -1,0 +1,32 @@
+import { css } from 'lit'
+
+export const styles = css`
+  @layer base, reset, transitional-styles;
+  @layer base {
+    :host {
+      display: inline-block;
+      max-width: 100%;
+    }
+
+    .chips {
+      display: flex;
+      flex-wrap: wrap;
+      gap: 0.5rem;
+    }
+
+    .collapse {
+      overflow: auto;
+      scroll-snap-type: both mandatory;
+      overscroll-behavior-x: contain;
+      flex-wrap: nowrap;
+      padding-bottom: 0.5rem;
+      scrollbar-color: rgba(0, 0, 0, 0.5) transparent;
+      scrollbar-width: thin;
+    }
+
+    ::slotted(*) {
+      scroll-snap-align: start;
+      scroll-snap-stop: always;
+    }
+  }
+`