+++ libs/react-charts/.storybook/preview.js
@@ -1,7 +1,0 @@
-import { parameters as baseParameters } from '../../../.storybook/preview'
-// eslint-disable-next-line import/no-webpack-loader-syntax
-import '!style-loader!css-loader!postcss-loader!sass-loader!./preview.scss'
-
-export const parameters = {
-  ...baseParameters,
-}